requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
pandas==2.2.0
python-dotenv==1.0.1
